Change Travis CI to 'language: generic'*
The 'make github' target has been broken for a long time, but due to a
lack of attention to this project it was never properly fixed. I want to
fix that now, but half expect to need several steps (commits) in order
to actually get things working again :-).

The actual breakage is a 128 exit code somewhere in the makefile, which
indicates a command not found error. While looking at the Travis CI logs
I noticed that two Python environments are being used rather
inconsistently. This might be the root cause of the issue or it might
just be an additional confusing aspect...

Given that the makefile unconditionally (re)creates its own Python
virtual environment, I've decided to configure Travis CI to opt out
of the Travis CI provided virtual environment.